Matching2 Part1: general additions, corrections, and sorting.
